test: FeedbackAppender append method test

+++ src/main/java/com/roundtable/roundtable/domain/feedback/PredefinedFeedbackRepository.java
@@ -3,6 +3,6 @@ package com.roundtable.roundtable.domain.feedback;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PredefinedFeedbackRepository extends JpaRepository<PredefinedFeedback, Long> {
-    List<PredefinedFeedback> findByIdIn(List<Long> id);
+public interface PredefinedFeedbackRepository extends JpaRepository<PredefinedFeedback, Integer> {
+    List<PredefinedFeedback> findByIdIn(List<Integer> id);
 }
